% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_project.R
\name{create_project}
\alias{create_project}
\title{Create a new D2SB project from the standard template}
\usage{
create_project(project_name, path = NULL, open = interactive(), ...)
}
\arguments{
\item{project_name}{Name of the new project (used for folder name, project.yml, and .Rproj file).}

\item{path}{Directory where the project should be created. If `NULL`,
the function will create a folder named `project_name` in the current working directory.}

\item{open}{Logical; whether to open the new project in RStudio.}

\item{...}{Named parameters to fill into `project.yml` template
(e.g., title, analyst, pi, status, etc.)}
}
\description{
Create a new D2SB project from the standard template
}
