% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/move_to_dissemination.R
\name{move_to_dissemination}
\alias{move_to_dissemination}
\title{Move a rendered D2SB Quarto report into the SharePoint Dissemination folder}
\usage{
move_to_dissemination(path_from, subdir_to, overwrite = FALSE, ...)
}
\arguments{
\item{path_from}{Path to the report file that exists (e.g. "Reports/eda.html").}

\item{subdir_to}{which subdirectory on sharepoint to copy to}

\item{overwrite}{whether to overwrite or not}

\item{...}{Named parameters to substitute into the template (title, author, etc.).}
}
\description{
Move a rendered D2SB Quarto report into the SharePoint Dissemination folder
}
