% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project_meta_data.R
\name{set_project_data}
\alias{set_project_data}
\title{Set data for project}
\usage{
set_project_data(param, value, overwrite = FALSE)
}
\arguments{
\item{param}{Project parameter to be written}

\item{value}{Value to assign to project parameter named param}

\item{overwrite}{If found, should parameter be over-written}
}
\value{
a named character vector with the requested param, or all project data
}
\description{
Allows you to write misc project data parameters
}
