% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project_meta_data.R
\name{sp_path}
\alias{sp_path}
\title{get the path to where data live on sharepoint for this project}
\usage{
sp_path(file_name = "")
}
\arguments{
\item{file_name}{optionally returns full file name for a specific file}
}
\value{
a director or file_name path that will return the proper
local location of a file or directory for the project.
}
\description{
use instead of `here` or `setwd` to get data location from sharepoint
for D2SB projects.
}
